refactor: refactor supporting functions for ec tests
